refactor: changed data type from 'int' to 'double'

+++ src/main/java/axi/practice/data_generation_reports/dto/group_request_stat/GroupRequestStatDto.java
@@ -15,6 +15,6 @@ public class GroupRequestStatDto {
     private UUID id;
     private String host;
     private String path;
-    private int avgHeaders;
-    private int avgQueryParams;
+    private double avgHeaders;
+    private double avgQueryParams;
 }
